ERL: Make a few internal to loader types public
This is in preparation for moving some of the currently internal API
public. Mostly cause it's probably nicer to do so
